Refactor common/index
