+++ apps/charterafrica/src/lib/data/common/processPageExplainers.js
@@ -1,0 +1,16 @@
+async function processPageExplainers(page, api, context) {
+  const collection = await api.getCollection("explainers", context);
+  const explainers = collection.docs || null;
+  const { title, blocks } = page;
+  if (explainers?.length) {
+    blocks.push({
+      slug: "explainers",
+      title,
+      explainers,
+    });
+  }
+
+  return page;
+}
+
+export default processPageExplainers;
